requirements.txt
torch
torchvision
pennylane
matplotlib
Pillow